Move all contact info struct in one file
