+++ include/dem/contact_info.h
@@ -14,20 +14,29 @@
  * ---------------------------------------------------------------------
  */
 
-#ifndef lethe_particle_wall_contact_info_h
-#define lethe_particle_wall_contact_info_h
+#ifndef lethe_contact_info_h
+#define lethe_contact_info_h
 
 #include <deal.II/base/point.h>
 #include <deal.II/base/tensor.h>
 
 #include <deal.II/particles/particle_iterator.h>
 
-/**
- * @brief This struct handles the information related to the calculation of the
- * particle-wall contact force
- */
-
 using namespace dealii;
+
+/**
+ * @brief Handle the information related to the calculation of the
+ * particle-particle contact force. Notably it is responsible for storing
+ * information that has to be preserved over multiple iterations of a contact,
+ * namely everything related to tangential overlaps
+ */
+template <int dim>
+struct particle_particle_contact_info
+{
+  Particles::ParticleIterator<dim> particle_one;
+  Particles::ParticleIterator<dim> particle_two;
+  Tensor<1, 3>                     tangential_overlap;
+};
 
 template <int dim>
 class particle_wall_contact_info
@@ -39,13 +48,16 @@ public:
    * boundary id. This is the commonly used constructor since it houses all the
    * information required to perform the contact calculation.
    *
-   * @param particle The iterator to the particle in contact with the wall
-   * @param normal_vector The outward pointing normal vector on the wall
-   * @param point_on_boundary A point that lies on the face
-   * @param boundary_id The boundary id. This id corresponds to the number attributed to the boundary condition.
+   * @param particle The iterator to the particle in contact with the wall.
+   * @param normal_vector The outward pointing normal vector on the wall.
+   * @param point_on_boundary A point that lies on the face.
+   * @param boundary_id The boundary id. This id corresponds to the number
+   * attributed to the boundary condition.
    *
+   * TODO: This should be a struct and normal_overlap, normal_relative_velocity
+   * and tangential_relative_velocity should be removed and be calculated on the
+   * fly as done for pp forces with particle_particle_contact_info
    */
-
   particle_wall_contact_info(const Particles::ParticleIterator<dim> &particle,
                              const Tensor<1, 3>       &normal_vector,
                              const Point<3>           &point_on_boundary,
@@ -87,6 +99,29 @@ public:
   double                           normal_relative_velocity;
   Tensor<1, 3>                     tangential_overlap;
   Tensor<1, 3>                     tangential_relative_velocity;
+};
+
+/**
+ * @brief Handle information related to the calculation of the particle-line
+ * contact forces.
+ */
+template <int dim>
+struct particle_line_contact_info
+{
+  Particles::ParticleIterator<dim> particle;
+  Point<3>                         point_one;
+  Point<3>                         point_two;
+};
+
+/**
+ * @brief Handle information related to the calculation of the particle-point
+ * contact forces.
+ */
+template <int dim>
+struct particle_point_contact_info
+{
+  Particles::ParticleIterator<dim> particle;
+  Point<3>                         point;
 };
 
 #endif
