+++ include/dem/force_chains_visualization.h
@@ -25,7 +25,6 @@
 #include <dem/dem_contact_manager.h>
 #include <dem/dem_solver_parameters.h>
 #include <dem/particle_particle_contact_force.h>
-#include <dem/particle_particle_contact_info.h>
 #include <dem/rolling_resistance_torque_models.h>
 
 #include <deal.II/particles/particle_handler.h>
