+++ include/dem/grid_motion.h
@@ -17,9 +17,9 @@
 #ifndef lethe_grid_motion_h
 #define lethe_grid_motion_h
 
+#include <dem/contact_info.h>
 #include <dem/data_containers.h>
 #include <dem/dem_solver_parameters.h>
-#include <dem/particle_wall_contact_info.h>
 
 #include <deal.II/distributed/tria.h>
