+++ include/dem/particle_particle_contact_force.h
@@ -20,10 +20,10 @@
 #include <core/auxiliary_math_functions.h>
 #include <core/dem_properties.h>
 
+#include <dem/contact_info.h>
 #include <dem/data_containers.h>
 #include <dem/dem_contact_manager.h>
 #include <dem/dem_solver_parameters.h>
-#include <dem/particle_particle_contact_info.h>
 #include <dem/rolling_resistance_torque_models.h>
 
 #include <deal.II/particles/particle_handler.h>
