--- include/dem/particle_point_line_contact_info.h
@@ -1,49 +1,0 @@
-/* ---------------------------------------------------------------------
- *
- * Copyright (C) 2019 - 2024 by the Lethe authors
- *
- * This file is part of the Lethe library
- *
- * The Lethe library is free software; you can use it, redistribute
- * it, and/or modify it under the terms of the GNU Lesser General
- * Public License as published by the Free Software Foundation; either
- * version 2.1 of the License, or (at your option) any later version.
- * The full text of the license can be found in the file LICENSE at
- * the top level of the Lethe distribution.
- *
- * ---------------------------------------------------------------------
- */
-
-#ifndef lethe_particle_point_line_contact_info_struct_h
-#define lethe_particle_point_line_contact_info_struct_h
-
-#include <deal.II/base/point.h>
-#include <deal.II/base/tensor.h>
-
-#include <deal.II/particles/particle_iterator.h>
-using namespace dealii;
-
-/**
- * @brief Handle information related to the calculation of the particle-line contact forces.
- */
-
-template <int dim>
-struct particle_line_contact_info
-{
-  Particles::ParticleIterator<dim> particle;
-  Point<3>                         point_one;
-  Point<3>                         point_two;
-};
-
-/**
- * @brief Handle information related to the calculation of the particle-point
- * contact forces.
- */
-template <int dim>
-struct particle_point_contact_info
-{
-  Particles::ParticleIterator<dim> particle;
-  Point<3>                         point;
-};
-
-#endif
