+++ include/dem/particle_wall_contact_force.h
@@ -21,9 +21,9 @@
 #include <core/dem_properties.h>
 #include <core/serial_solid.h>
 
+#include <dem/contact_info.h>
 #include <dem/data_containers.h>
 #include <dem/dem_solver_parameters.h>
-#include <dem/particle_wall_contact_info.h>
 
 #include <boost/math/special_functions.hpp>
 #include <boost/range/adaptor/map.hpp>
