+++ include/dem/particle_wall_dmt_force.h
@@ -20,7 +20,6 @@
 #include <core/dem_properties.h>
 
 #include <dem/dem_solver_parameters.h>
-#include <dem/particle_wall_contact_info.h>
 #include <dem/particle_wall_nonlinear_force.h>
 
 #include <deal.II/particles/particle.h>
