+++ include/fem-dem/ib_particles_dem.h
@@ -15,22 +15,20 @@
  *
  */
 
+#ifndef lethe_ib_particles_dem_h
+#define lethe_ib_particles_dem_h
+
 #include <core/ib_particle.h>
 #include <core/ib_stencil.h>
 #include <core/lethe_grid_tools.h>
 
 #include <dem/particle_particle_contact_force.h>
-#include <dem/particle_particle_contact_info.h>
 #include <dem/particle_wall_contact_force.h>
-#include <dem/particle_wall_contact_info.h>
 
 #include <deal.II/base/tensor.h>
 
 #include <deal.II/dofs/dof_handler.h>
 #include <deal.II/dofs/dof_tools.h>
-
-#ifndef lethe_ib_particles_dem_h
-#  define lethe_ib_particles_dem_h
 
 using namespace dealii;
