+++ source/dem/CMakeLists.txt
@@ -50,6 +50,7 @@ add_library(lethe-dem
   # Headers
   ../../include/dem/adaptive_sparse_contacts.h
   ../../include/dem/boundary_cells_info_struct.h
+  ../../include/dem/contact_info.h
   ../../include/dem/contact_type.h
   ../../include/dem/data_containers.h
   ../../include/dem/dem.h
@@ -76,15 +77,12 @@ add_library(lethe-dem
   ../../include/dem/output_force_torque_calculation.h
   ../../include/dem/particle_particle_broad_search.h
   ../../include/dem/particle_particle_contact_force.h
-  ../../include/dem/particle_particle_contact_info.h
   ../../include/dem/particle_particle_fine_search.h
   ../../include/dem/particle_point_line_broad_search.h
   ../../include/dem/particle_point_line_contact_force.h
-  ../../include/dem/particle_point_line_contact_info.h
   ../../include/dem/particle_point_line_fine_search.h
   ../../include/dem/particle_wall_broad_search.h
   ../../include/dem/particle_wall_contact_force.h
-  ../../include/dem/particle_wall_contact_info.h
   ../../include/dem/particle_wall_dmt_force.h
   ../../include/dem/particle_wall_fine_search.h
   ../../include/dem/particle_wall_jkr_force.h
