+++ tests/dem/particle_point_contact.cc
@@ -36,6 +36,7 @@
 // Lethe
 #include <core/dem_properties.h>
 
+#include <dem/contact_info.h>
 #include <dem/dem_solver_parameters.h>
 #include <dem/find_boundary_cells_information.h>
 #include <dem/particle_point_line_broad_search.h>
@@ -128,7 +129,7 @@ test()
     contact_candidates;
 
   // Particle-point fine search
-  std::unordered_map<unsigned int, particle_point_line_contact_info<dim>>
+  std::unordered_map<unsigned int, particle_point_contact_info<dim>>
     contact_information;
 
   ParticlePointLineForce<dim>   force_object;
